+++ config/dev.exs
@@ -1,3 +1,7 @@
 use Mix.Config
 
-config :ex_cldr, MyApp.Gettext, default_locale: "en"
+config :ex_cldr,
+  default_backend: MyApp.Cldr
+
+config :ex_cldr, MyApp.Gettext,
+  default_locale: "en"