+++ lib/cldr/cldr_chars_protocol.ex
@@ -21,3 +21,16 @@ defimpl Cldr.Chars, for: Cldr.LanguageTag do
     Cldr.LanguageTag.to_string(language_tag)
   end
 end
+
+defimpl Cldr.Chars, for: BitString do
+  def to_string(term) when is_binary(term) do
+    term
+  end
+
+  def to_string(term) do
+    raise Protocol.UndefinedError,
+      protocol: @protocol,
+      value: term,
+      description: "cannot convert a bitstring to a string"
+  end
+end