+++ test/plural_rules/decimal_plural_rule_test.exs
@@ -1,5 +1,5 @@
 defmodule Decimal.PluralRule.Test do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   @moduletag :slow
   @sample_types [:decimal]