+++ test/units_test.exs
@@ -1,5 +1,5 @@
 defmodule CldrUnitsTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
 
   test "parse unit expression with power" do
     assert Cldr.Unit.Parser.parse("ft2m^3 * 43560") == ["*", ["^", "ft2m", 3], 43560]